requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
colorama>=0.4.6
